pyproject.toml
[project]
name = "music-discord-bot"
version = "0.1.0"
description = "A Discord music bot using discord.py and Wavelink"
authors = [{ name = "Example" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "discord.py>=2.3",
  "wavelink>=3.2",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["black>=24.0", "ruff>=0.4"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
